Add general method to External login, now FaceBook and Google works
