+++ OnlineLibrary/Models/UserAuthModel.cs
@@ -11,4 +11,41 @@ namespace OnlineLibrary.Models
         public UserLoginModel UserLoginModel { get; set; }
         public UserRegisterModel UserRegisterModel { get; set; }
     }
+
+    public class UserLoginModel
+    {
+        [Required]
+        [EmailAddress]
+        [Display(Name = "Email")]
+        public string Email { get; set; }
+
+        [Required]
+        [DataType(DataType.Password)]
+        [Display(Name = "Password")]
+        public string Password { get; set; }
+    }
+
+    public class UserRegisterModel
+    {
+        [Required]
+        [EmailAddress]
+        [Display(Name = "Email")]
+        public string Email { get; set; }
+
+        [Required]
+        [DataType(DataType.Password)]
+        [Display(Name = "Password")]
+        public string Password { get; set; }
+
+        [Required(ErrorMessage = "Password confirm field is required")]
+        [Compare("Password", ErrorMessage = "Passwords do not match ")]
+        [Display(Name = "Confirm password")]
+        [DataType(DataType.Password)]
+        public string PasswordConfirm { get; set; }
+    }
+
+    public class ExternalLoginListViewModel
+    {
+        public string ReturnUrl { get; set; }
+    }
 }
