--- OnlineLibrary/Models/UserRegisterModel.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Web;
-
-namespace OnlineLibrary.Models
-{
-    public class UserRegisterModel
-    {
-        [Required]
-        [EmailAddress]
-        [Display(Name = "Email")]
-        public string Email { get; set; }
-
-        [Required]
-        [DataType(DataType.Password)]
-        [Display(Name = "Password")]
-        public string Password { get; set; }
-
-        [Required(ErrorMessage = "Password confirm field is required")]
-        [Compare("Password", ErrorMessage = "Passwords do not match ")]
-        [Display(Name = "Confirm password")]
-        [DataType(DataType.Password)]
-        public string PasswordConfirm { get; set; }
-    }
-}
