+++ OnlineLibrary/Web.config
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <!--
   For more information on how to configure your ASP.NET application, please visit
   https://go.microsoft.com/fwlink/?LinkId=169433
@@ -20,15 +20,13 @@
     <!-- adjundiet: Add Connection string to local db -->
     <!--<add name="DBConnection" connectionString="data source=(localdb)\MSSQLLocalDB;Initial Catalog=userstore.mdf;Integrated Security=True;"
          providerName="System.Data.SqlClient"/> -->
-    <add name="onlinelibrary2018" 
-         providerName="System.Data.SqlClient" 
-         connectionString="data source=(localdb)\MSSQLLocalDB;Initial Catalog=onlinelibrary2018.mdf;Integrated Security=True;" />
+    <add name="onlinelibrary2018" providerName="System.Data.SqlClient" connectionString="data source=(localdb)\MSSQLLocalDB;Initial Catalog=onlinelibrary2018.mdf;Integrated Security=True;" />
   </connectionStrings>
   
   
   <!-- adjundiet:__________________________________ -->
   <system.web>
-        <customErrors mode="Off"/>
+        <customErrors mode="Off" />
 
     <compilation debug="true" targetFramework="4.6.1" />
     <httpRuntime targetFramework="4.6.1" />
