+++ OnlineLibrary/packages.config
@@ -15,6 +15,7 @@
   <package id="Microsoft.Owin.Host.SystemWeb" version="4.0.0" targetFramework="net461" />
   <package id="Microsoft.Owin.Security" version="4.0.0" targetFramework="net461" />
   <package id="Microsoft.Owin.Security.Cookies" version="4.0.0" targetFramework="net461" />
+  <package id="Microsoft.Owin.Security.Facebook" version="4.0.0" targetFramework="net461" />
   <package id="Microsoft.Owin.Security.Google" version="4.0.0" targetFramework="net461" />
   <package id="Microsoft.Owin.Security.OAuth" version="4.0.0" targetFramework="net461" />
   <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net461" />
